fix: combine all pipeline file
