--- shell/app/modules/application/common/yml-editor/chart/yml-chart.scss
@@ -1,24 +1,0 @@
-.yml-svg-container {
-  position: relative;
-  min-width: 100%;
-  min-height: 100%;
-  background-image: url(/images/editor-background.png);
-  background-repeat: repeat;
-  background-size: 100%;
-
-  &.editing {
-    min-height: calc(100% - 210px);
-  }
-
-  .svg-model-node-carrier {
-    overflow: visible;
-
-    &.hidden {
-      visibility: hidden;
-    }
-
-    &.show {
-      visibility: visible;
-    }
-  }
-}
