--- shell/app/modules/application/common/yml-editor/pipeline-yml/pipeline-node-drawer.scss
@@ -1,5 +1,0 @@
-.yml-node-drawer {
-  .ant-drawer-body {
-    overflow: auto;
-  }
-}
