--- shell/app/modules/application/common/yml-editor/yml-node/add-row.scss
@@ -1,18 +1,0 @@
-.yml-chart-node {
-  &.add-line {
-    width: 100%;
-    height: 100%;
-    color: $lightgray;
-    font-size: 24px;
-    line-height: 14px;
-    text-align: center;
-    border: 1px solid $lightgray;
-    border-radius: 100%;
-    cursor: pointer;
-
-    &:hover {
-      color: $primary;
-      border: 1px solid $primary;
-    }
-  }
-}
