--- shell/app/modules/application/common/yml-editor/yml-node/start-node.scss
@@ -1,9 +1,0 @@
-.yml-chart-node {
-  &.start-node {
-    width: 100%;
-    height: 100%;
-    background-color: $white;
-    border: 1px solid $color-border;
-    border-radius: 100px;
-  }
-}
