+++ shell/app/yml-chart/common/action-config-form.tsx
@@ -10,27 +10,25 @@
 //
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
-
 import * as React from 'react';
-import { Icon as CustomIcon } from 'common';
-import i18n from 'i18n';
-import './add-node.scss';
+import DiceConfigPage from 'app/config-page/index';
 
-export interface IProps{
-  data: Obj;
-  onClickNode: (data: any, arg?: any) => void;
+interface IProps{
+  nodeData: null | IStageTask;
+  editing: boolean;
+  isCreate?: boolean;
+  otherTaskAlias?: string[];
+  chosenActionName: string;
+  chosenAction:any;
+  onSubmit?: (options: any) => void;
 }
-const noop = () => {};
-export const AddNode = (props: IProps) => {
-  const { data, onClickNode = noop } = props;
 
-  const onClick = () => {
-    onClickNode(data);
-  };
-  return (
-    <div className='yml-chart-node add-node' onClick={onClick}>
-      <CustomIcon type='tj1' className='add-icon mb12' />
-      <span className='add-node-txt'>{i18n.t('add {name}', { name: i18n.t('node') })}</span>
-    </div>
-  );
+const ActionConfigForm = (props: IProps) => {
+  const { chosenActionName } = props;
+  const inParams = {
+    actionData: props.nodeData
+  }
+  return <DiceConfigPage inParams={inParams} showLoading={false} scenarioType='action' scenarioKey={chosenActionName} customProps={{ actionForm: props }} />;
 };
+
+export default ActionConfigForm;
