+++ shell/app/yml-chart/nodes/start-node.tsx
@@ -13,11 +13,12 @@
 
 import * as React from 'react';
 import i18n from 'i18n';
+import classnames from 'classnames';
 import './start-node.scss';
 
 export interface IProps{
-  data: Obj;
-  onClickNode: (data: any, arg?: any) => void;
+  data?: Obj;
+  onClickNode?: (data: any, arg?: any) => void;
   disabled?: boolean;
 }
 
@@ -30,8 +31,15 @@ export const StartNode = (props: IProps) => {
     !disabled && onClickNode(data);
   };
 
+  const classes = classnames({
+    'yml-chart-node': true,
+    'start-node': true,
+    'center-flex-box': true,
+    'hover-active': !disabled,
+  });
+
 
   return (
-    <div className='yml-chart-node start-node pointer center-flex-box hover-active' onClick={onClick}>{ disabled ? '' : i18n.t('project:params configuration')}</div>
+    <div className={classes} onClick={onClick}>{ disabled ? '' : i18n.t('project:params configuration')}</div>
   );
 };
